move Room to websocket.models package
