+++ app/src/main/java/ru/mkirilkin/doodlekong/ui/setup/select_room/SelectRoomViewModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
-import ru.mkirilkin.doodlekong.data.remote.websocket.Room
+import ru.mkirilkin.doodlekong.data.remote.websocket.models.Room
 import ru.mkirilkin.doodlekong.repository.SetupRepository
 import ru.mkirilkin.doodlekong.util.DispatcherProvider
 import ru.mkirilkin.doodlekong.util.Resource
